--- core/cas-server-core-services-api/src/main/java/org/apereo/cas/services/RegisteredServicesEventListenerAsync.java
@@ -1,44 +1,0 @@
-package org.apereo.cas.services;
-
-import org.apereo.cas.support.events.service.CasRegisteredServiceExpiredEvent;
-import org.apereo.cas.support.events.service.CasRegisteredServicesRefreshEvent;
-import org.apereo.cas.util.spring.CasEventListener;
-import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
-
-/**
- * Interface for {@code RegisteredServicesEventListener} to allow spring {@code @Async} support to use JDK proxy.
- * @author Hal Deadman
- * @since 6.5.0
- */
-public interface RegisteredServicesEventListenerAsync extends CasEventListener  {
-
-    /**
-     * Handle services manager refresh event.
-     *
-     * @param event the event
-     */
-    @EventListener
-    @Async
-    void handleRefreshEvent(CasRegisteredServicesRefreshEvent event);
-
-    /**
-     * Handle environment change event.
-     *
-     * @param event the event
-     */
-    @EventListener
-    @Async
-    void handleEnvironmentChangeEvent(EnvironmentChangeEvent event);
-
-    /**
-     * Handle registered service expired event.
-     *
-     * @param event the event
-     */
-    @EventListener
-    @Async
-    void handleRegisteredServiceExpiredEvent(CasRegisteredServiceExpiredEvent event);
-
-}
