+++ core/cas-server-core-services-api/src/main/java/org/apereo/cas/services/RegisteredServicesEventListenerImpl.java
@@ -1,0 +1,83 @@
+package org.apereo.cas.services;
+
+import org.apereo.cas.configuration.CasConfigurationProperties;
+import org.apereo.cas.notifications.CommunicationsManager;
+import org.apereo.cas.notifications.mail.EmailMessageBodyBuilder;
+import org.apereo.cas.support.events.service.CasRegisteredServiceExpiredEvent;
+import org.apereo.cas.support.events.service.CasRegisteredServicesRefreshEvent;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import lombok.val;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.scheduling.annotation.Async;
+
+import java.util.Map;
+
+/**
+ * This is {@link RegisteredServicesEventListenerImpl}.
+ *
+ * @author Misagh Moayyed
+ * @since 5.1.0
+ */
+@RequiredArgsConstructor
+@Slf4j
+public class RegisteredServicesEventListenerImpl implements RegisteredServicesEventListener {
+    private final ServicesManager servicesManager;
+
+    private final CasConfigurationProperties casProperties;
+
+    private final CommunicationsManager communicationsManager;
+
+    @Override
+    @EventListener
+    @Async
+    public void handleRefreshEvent(final CasRegisteredServicesRefreshEvent event) {
+        servicesManager.load();
+    }
+
+    @Override
+    @EventListener
+    @Async
+    public void handleEnvironmentChangeEvent(final EnvironmentChangeEvent event) {
+        servicesManager.load();
+    }
+    
+    @Override
+    @EventListener
+    @Async
+    public void handleRegisteredServiceExpiredEvent(final CasRegisteredServiceExpiredEvent event) {
+        val registeredService = event.getRegisteredService();
+        val contacts = registeredService.getContacts();
+        val serviceRegistry = casProperties.getServiceRegistry();
+        val serviceName = StringUtils.defaultIfBlank(registeredService.getName(), registeredService.getServiceId());
+        if (contacts == null || contacts.isEmpty()) {
+            LOGGER.debug("No contacts are defined to be notified for policy changes to service [{}]", serviceName);
+            return;
+        }
+
+        val logMessage = String.format("Sending notification to [{}] as service [{}] is %s from registry", event.isDeleted() ? "deleted" : "expired");
+        LOGGER.info(logMessage, contacts, serviceName);
+
+        communicationsManager.validate();
+        if (communicationsManager.isMailSenderDefined()) {
+            val mail = serviceRegistry.getMail();
+            val body = EmailMessageBodyBuilder.builder().properties(mail)
+                .parameters(Map.of("service", serviceName)).build().produce();
+            contacts
+                .stream()
+                .filter(c -> StringUtils.isNotBlank(c.getEmail()))
+                .forEach(c -> communicationsManager.email(mail, c.getEmail(), body));
+        }
+        if (communicationsManager.isSmsSenderDefined()) {
+            val sms = serviceRegistry.getSms();
+            val message = sms.getFormattedText(serviceName);
+            contacts
+                .stream()
+                .filter(c -> StringUtils.isNotBlank(c.getPhone()))
+                .forEach(c -> communicationsManager.sms(sms.getFrom(), c.getPhone(), message));
+        }
+    }
+}
